git_ranger/requirements.txt
certifi==2018.10.15
chardet==3.0.4
Deprecated==1.2.4
Django==2.1.3
django-prometheus==1.0.15
django-reset-migrations==0.3.1
djangorestframework==3.9.0
idna==2.7
prometheus-client==0.4.2
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
PyGithub==1.43.3
PyJWT==1.6.4
pytz==2018.7
requests==2.20.1
urllib3==1.24.1
whitenoise==4.1.2
wrapt==1.10.11
